Creating AdvancedSearchManager to handle SearchAPI integration
